+++ lib/ng-vcl/src/embedded-input-group/interfaces.ts
@@ -1,10 +1,10 @@
 import { InjectionToken } from '@angular/core';
 
-export interface EmbeddedInputGroupHost {
+export interface EmbeddedInputGroup {
   disabled: boolean;
   prepIcon?: string;
   appIcon?: string;
   hasAppendedItem: boolean;
 }
 
-export const EMBEDDED_INPUT_GROUP_TOKEN = new InjectionToken<EmbeddedInputGroupHost>('embedded-input-group');
+export const EMBEDDED_INPUT_GROUP_TOKEN = new InjectionToken<EmbeddedInputGroup>('embedded-input-group');